:sparkles: User can now switch between metric/american units, fix header into proper navbar
-This is a big one, had to move my HOC from Main to App in order to
facilitate easier handling of the unit conversion. Also had to make
a separate API request for the US units. The time added is negligble
compared to what's gained from seamless unit switch. The main reason
why the previous conversion method was scrapped was that it just made
more sense to switch the entire system instead of mixing SI units and
Imperial. This also saved me time on parsing and converting the summary
string responses.